+++ pages/poetry/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Link from "next/link";
-import { List, FlexboxGrid, Placeholder } from "rsuite";
+import { List, FlexboxGrid } from "rsuite";
 
 import styles from "../../styles/Home.module.css";
 import "rsuite/dist/rsuite.min.css";
@@ -25,10 +25,11 @@ const entry: PoetryEntry[] = [
       "今年は夏が来るのが早すぎたので、思い出の中の夏を考えるのと暑さを体感するのが同時になってしまった。思い出と現実が混ざりそうなので、思ったままに書いてみた。",
   },
   {
-    date: "2022-07-10",
-    title: "未定",
-    link: "/poetry/2_unknown",
-    description: "何も決まっていないけど、とりあえず書いてみる。",
+    date: "2022-07-17",
+    title: "東京",
+    link: "/poetry/2_tokyo",
+    description:
+      "東京を歩きながら、生き方についてぼんやりと考えたことを表現してみた。",
   },
 ];
